Switched the set-up of the repository to define a cargo workspace. The functionalities placed in hydrophonitor-lib are used by hydrophonitor-cli. In the future, these can also be used by hydrophonitor-gui.
