+++ Cargo.toml
@@ -1,16 +1,16 @@
-[package]
-name = "hydrophonitor-cli"
-version = "0.1.0"
-edition = "2021"
+[workspace]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+members = [
+    "cli",
+    "lib",
+]
 
-[dependencies]
+[workspace.dependencies]
 anyhow = "1.0.70"
 clap = { version = "4.1.11", features = ["derive"] }
+clap-verbosity-flag = "2.1.0"
+env_logger = "0.10.1"
 hound = "3.5.0"
 indicatif = "0.17.3"
-walkdir = "2.3.3"
 log = "0.4"
-env_logger = "0.10.1"
-clap-verbosity-flag = "2.1.0"
+walkdir = "2.3.3"
