+++ cli/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "hydrophonitor-cli"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+hydrophonitor-lib = { path = "../lib" }
+anyhow.workspace = true
+clap.workspace = true
+clap-verbosity-flag.workspace = true
+env_logger.workspace = true
+hound.workspace = true
+indicatif.workspace = true
+log.workspace = true
+walkdir.workspace = true
