+++ lib/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "hydrophonitor-lib"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow.workspace = true
+hound.workspace = true
+indicatif.workspace = true
+log.workspace = true
+walkdir.workspace = true
